requirements.txt
anthropic>=0.21.0
beautifulsoup4>=4.12.0
chromadb>=0.4.0
fastapi>=0.110.0
fuzzywuzzy>=0.18.0
loguru>=0.7.3
nbconvert>=7.16.6
openai>=1.82.0
pocketflow>=0.0.1
python-dotenv>=1.0.1
python-Levenshtein>=0.21.0
pyyaml>=6.0
requests>=2.31.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
streamlit>=1.32.0
streamlit-extras>=0.3.5
typing_extensions>=4.0.0
uvicorn>=0.27.1
